filesystem structure (#2)
* Add contribuiting file

Created the contribution file to be used to explain the guidelines for
completing a proper 'pull request'.
Inside is a comment with a link to a github help page useful for
understanding how to build the file.

* create fs structure

created filesystem structure with README files inside the directories to keep
the strucutre consistent

* create gitignore

the python cache directories and the log files added to the new
gitignore file

* create docker folder and folder for logger service

added a docker folder where to put all the docker files (Dockerfile and
docker-compose) to create images for all the necessary services.